Added photos route and refactored some of the queries from other controllers into photos controller
